Add missing trailing slash
By looking at the top of the document "<!DOCTYPE html>" it seems we
don't need trailing slash for self-closing tags but maybe the parser
needs (?). This seems to be super safe and it would be nice to see if
it fixes the "fails to build a proper html tag, unclosed link tag" problem.

- For consistency (in other lines we use it)
- For XHTML-compatibility

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: I00bf34104a5b56d094196cb2b912e1dc308b6253