fix(android): Restore progress bar
On Android the document loading progress bar had regressed so it no
longer moved along as the document loaded. Instead, it stayed at 0%
until the document loaded, when it disappeared without having moved at
all...

...this was a known regression from
https://github.com/CollaboraOnline/online/pull/8957 -- known as in
"known before we merged it". We decided background save was important
enough that breaking the -- poor as discussed in the next paragraphs --
progress bar was worth merging it, and we would fix the regression
later. This is that fix.

The fix here still freezes for a surprisingly long time on "empty" -
enough that I wonder if there's something freezing the UI thread during
load - however it is at a similar level to before the regression and is
a definite improvement over the current state...

...I will continue looking to make the progress bar "look alive" earlier
during document load...

Fixes: https://github.com/CollaboraOnline/online/issues/9694
Change-Id: Idc30105cdd2f650bc5d39deae3437aa96bfe32d1
Signed-off-by: Skyler Grey <skyler.grey@collabora.com>