fix(calc): Restore inverted document on load
Previously, we didn't have a way to restore the document background
color on load for Calc, leading to a document that was inverted
according to our LocalStorage and Core, but which didn't have the
background to show it. This led to text rendering issues, as well as
needing multiple presses to switch into invert background mode
properly...

...in this change, I've simplified the calc background code to rely on
our existing css data-bg-theme property, and made it consistently set
the background color to --color-background-document ... it's possible
this causes problems with changing themes alongside
LOK_CALLBACK_DOCUMENT_BACKGROUND_COLOR, but I'm confident this is an
improvement over the previous behavior as that should also have been
true of it...

...this brings Calc invert backgrounds to parity with the other apps, in
terms of saving/loading them properly...

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I45b2fe174f1e7d0ed8fb6bc582a5932c86d4d3d1
Fixes: https://github.com/CollaboraOnline/online/issues/9538