if creating a user ns worked, but mount doesn't then map orig uid
if, like with apparmor, user namespaces are possible, but mounts are not
(not even the initial "none" mount to not propogate mounts to the host
env) then we are in a namespace, but not seen in our env as any
particular user.

that leads to worrying warnings about "incorrect user-name" like:

[ coolwsd ] DBG  Move into user namespace as uid 0| wsd/COOLWSD.cpp:1965
[ coolwsd ] ERR  enterMountingNS, root mount failed: Inappropriate ioctl for device| common/JailUtil.cpp:78
[ coolwsd ] ERR  creating usernamespace for mount user failed.| wsd/COOLWSD.cpp:1969
[ coolwsd ] INF  Cleaning up childroot directory [/opt/cool/child-roots/].| common/JailUtil.cpp:307
[ coolwsd ] TRC  Checking pid for jail 3192148 /opt/cool/child-roots/| common/JailUtil.cpp:335
[ coolwsd ] DBG  Removing [/opt/cool/child-roots/3192148-91211bec/tmp] recursively.| common/FileUtil.cpp:210
[ coolwsd ] DBG  Removing [/opt/cool/child-roots/3192148-91211bec/linkable] recursively.| common/FileUtil.cpp>
Security: coolmount incorrect user-name, other than 'cool'
Aborting.

so on namespace success, but initial mount failure, map the namespace
user the original uid so coolmount is happy to continue

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: Ia4b5c79ae60c2c588a645850924728951d64446b