feat: Changes to UnitPerf measuring
Previously all metrics were being recorded into seperate CSV's and without the git commit hash, date, type of the test . Also CPU usage and stress run timings were an aggregate of the setup + load + edit phases.

Now each metric is recorded into a summary CSV with the git commit hash as a primary key followed by date, type of test (writer, calc etc..), the phase (setup,load,edit) , the metric and of course the value. This is to allow the use of a pivot table and allow for some nice charting.

Signed-off-by: elliotfreebairn <elliot.freebairn@collabora.com>
Change-Id: I868a0a1c91b30392ab002ce149cefae0dd3036cc