test: Make cypress fail on exception
In I636e07a93d2ed17d56b2a9fc11d66b7e4c3c19f3, Michael Meeks added a new
popup to show when we have an exception in an event. This is excellent,
and is catching errors. Unfortunately, the messages we get in our
Cypress tests when this fails are... not amazing.

They don't include the error, and instead the new popup just makes some
actions that we wanted to perform silently fail to happen. This causes
later assertion errors, but they are disconnected from the root cause -
the error. It's much neater to quit the test immediately, which we can
do by throwing the error.

Unfortunately, we can't immediately throw the error as we are inside an
iframe, stopping Cypress from seeing the error. Therefore, we have to
pass it up using a postMessage, which allows us to throw it in
`debug.html` (or `cypress-multiuser.html`) and fail the test

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Ia729489ba2adc3be0e9286f942d128acafd29858