slideshow: add more resolution adjustment options
Previously we topped out at 1920x1080 resolution, which can become
a problem for screens bigger than FHD, which are capable to show
a much higher resolution of the slides but are artificially limited
to a smaller resolution. This change adds 2560x1440 and 3840x2160
resolution. Ideally it would be to support resolution whatever the
screen supports, but as there are issues with that (sliedshow does
not show transitions correctly the second time) we add more options
as a stop gap solution.

Signed-off-by: Tomaž Vajngerl <tomaz.vajngerl@collabora.co.uk>
Change-Id: I9b132ad5ac897f956768c5d5a1e77fa18cb34246