Fix error on context change after mode switch
Fixes #9816

When we open classic/compact mode, then switch to notebookbar mode - the old component of top toolbar is not correctly removed. What leads to an error.

Steps to Reproduce:
1. Open any document, use compact mode
2. Switch to notebookbar mode
3. Change context, eg. select shape then select text

Don't use bind(this) so we successfully remove the same instance
of handler we inserted, we already pass "this" as 3rd param.
Notice we should also remove callback from app.events but it doesn't
have that function yet

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: Iaa8086b87f396c18f725666950683211238eaf41