fix(android): Correctly skip save if not modified
When the save command has a boolean value, it wants an integer (0 for
false or nonzero for true). Unfortunately, providing anything but an
integer is recognized as 'false', leading to some places in Android
incorrectly trying to save without changes.

In some document types, e.g. PDF, this causes a "could not save the
document" error.

I've corrected it and updated some of the documentation to make it a bit
clearer that it needs to be an integer.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: Iae06329982c3a044d9d35ad7d634990d6cb883e3