Keyboard shortcuts dialog: Improve formatting on copy shortcut
Before this commit the 2 shortcuts that trigger copy were placed
side by side with a comma in between, this was breaking the visual
table.

-> Add a break line so, it's easier to scan.

Note: the alternative would be using a rowspan="2" in the <tr> and
then add a new <tr> after that so, to achieve the same result but that
would break the search results (searching for copy would only show the
first shortcut). Conclusion: better to just add a </br>

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: Id1bd13dce0cacba7cdc38934f5b067fb26baf060