cool#9621: ClientRequestDispatcher::handleIncomingMessage() shutdown() POST FileRequestHandler->handleRequest() request served
We stumbled into a regression due to enforced keept connections (no shutdown),
leading to incomplete http requests (parseHeader).
The cause might be non-removed buffer content and has to be fully analyzed.

For now we shutdown POST fileserver request, mitigating the issue.
Thanx to Caolán finding the issue and Ashod discussing it further.

Signed-off-by: Sven Göthel <sven.gothel@collabora.com>
Change-Id: I83c09e0a8101931efece42e42e5a8e3abd7e696f