Fix build on async request usage
There was old check of success introduced in
commit 76fded3ef0065a04753422ba8b2910802943efcb
wsd: check for asyncRequest success for uploading

but in the meantime we have new callback for connection
fail what requires change from us:

commit f6ce37b579c1753e61a71788a78bb25d5d60eb1d
hook Session::asyncRequest up to use net::asyncConnect
calls _onConnectFail on async connect fail instead of returning false on
sync connect fail

So let's not use scoped callback trigger but rather do it manually
on returns and in the fail callback. Successfull one is already
done this way. Notice: on success we were firing it two times:
on Running state and Complete. Running doesn't happen if we fail
on connection...

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: I3815f0e236fca50c0ffd1be166f9eba3faa98529