input file to m4 macro through intermediate file, not parameter
This fixes a build issue where the parameter to m4 is too long as
it contains all the *.js files in the project. So instead of using
the parameter to pass those file, write them into an intermediate
file and then include the file in the m4 source instead.

Adapt patch to LibreOffice Online.

Signed-off-by: Tomaž Vajngerl <tomaz.vajngerl@collabora.co.uk>
Change-Id: Id31912720f684820683a8bec40ba451815e07dd1