refactor(clipboard): clean _onCommandResult wait
Previously we had bundled some of the fetching logic into waiting for a
command result, but that will make it hard to later use this for mobile
applications where fetching isn't used at all. Instead, we can clean
this up by moving out our fetching logic, and for ease-of-use of the new
API we can make a short function so we don't have to manually add
promises to the queue.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: If66e46459e497173a0e6cd114a45ec6972c698c0