slideshow: refactoring: move from SlideInfo to TransitionFilterInfo
This refactoring is a pre-step for being able to apply transition
filters to shapes/layers.
In fact a TransitionFilter node needs to pass to a transition filter
animation its own information not slide info.
So a TransitionFilterInfo class has been introduced.
This refactoring also removed the need to use
stringToTransitionSubTypeMap mapping in every sinlge transition, its
use is limited to the TransitionFilterInfo.fromSlideInfo function.

Adapt patch to LibreOffice Online.

Signed-off-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Change-Id: I9ed0357d9d25bf0791d15c6bd10ed9faa4fe4765