Refactor some elements in about dialog box.
We want to hide some elements outside of debugging mode, which is activated by triple click on about dialog.
That requires some elements to be created hidden by default.
This commit removes the "style="display:none"" code from HTML. Because that way of defaulting to hidden raises CSP issues.
Instead, we hide them right before we show the about dialog.

Also, there were duplicate IDs of elements in about dialog. Since we clone a dialog which already exists, all the IDs in that dialog are duplicated.
This commit adds "-cloned" suffix to 2 elements which have duplicate IDs. We probably need to check also other duplicates.

Adapt patch to LibreOffice Online.

Signed-off-by: Gökay Şatır <gokaysatir@collabora.com>
Change-Id: Idef8e842b2b2fd90a1639db63dd490d7718ff1c9