wsd: lock only after successful loading
For documents that are password-protected,
we fail to load the first time. Having taken
the lock before attempting to load, then
unlocking when we fail to load (due to the
missing password) is too disruptive.

Instead, we take the lock only after we
successfully load the document.

Adapt patch to LibreOffice Online.

Change-Id: Ie8ef0a10c16dd525d6bf8297b04bf35b64163535
Signed-off-by: Ashod Nakashian <ashod.nakashian@collabora.co.uk>