use jsdialog clearDialog directly to remove autocomplete popup
firing close, calls JSDialog::close which calls closePopover
which calls clearDialog before focusToLastElement, and clearDialog
removes the entry that focusToLastElement looks up to find the
previous thing that has focus, so it always ends up on the map
via fallback.

So explicitly call focusToLastElement first and then clearDialog, so we
get the focus returned to the text entry we want it to return to.

Adapt patch to LibreOffice Online.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: I0077216c9bebd780c9244ba99b32ba7d266c02fc