fix(ios): better approximate our version format
Previously we were using versions like "24.04.7" for release
"24.04.7.3". Unfortunately, this works poorly if we want to have a
patch release, as we can't reuse codes so we have to immediately go to
"24.04.8", even when releasing from a "24.04.7" tag.

Merging our year and major release into Apple's major release number
makes more sense with how we expect releases to change over time...

We also need to change how we generate the CFBundleVersion, as
otherwise it will be invalidly-long (the maximum length being 18
characters and "2404.7.202410021739" being a frustrating 19 characters,
for example

Adapt patch to LibreOffice Online and use a more focused on year and
month focused release naming.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I4aa9ab53b3a4d5786e04a90ab4c7c759d9606706